Implement program containers.

+++ libtuum/platform/src/platform.cpp
@@ -3,9 +3,9 @@
 
 using namespace std::chrono;
 
-size_t millis() {
+unsigned long int millis() {
   auto ep = high_resolution_clock::from_time_t(0);
   auto now = high_resolution_clock::now();
   auto msec = duration_cast<milliseconds>(now - ep).count();
-  return (size_t)msec;
+  return (unsigned long int)msec;
 }
